Criada a view especifica de Apartamento e Menu
